requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.6,<5.0
django-cors-headers>=3.7,<4.0
psycopg2>=2.8,<3.0
python-dotenv>=0.17,<1.0
celery>=5.1,<6.0
redis>=3.5,<4.0
requests>=2.25,<3.0
python-telegram-bot>=21.3